backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
groq>=0.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pillow>=10.0.0
python-telegram-bot>=20.0
instagrapi>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
